Create SetMembershipChecker extension
