+++ ChangeLog
@@ -10,12 +10,10 @@ Release date: TBA
 ..
   Put new features here and also in 'doc/whatsnew/2.11.rst'
 
-* ``CodeStyleChecker``
+* Added new extension ``SetMembershipChecker`` with ``use-set-for-membership`` check:
+  Emitted when using an in-place defined ``list`` or ``tuple`` to do a membership test. ``sets`` are better optimized for that.
 
-  * Added ``use-set-for-membership``: Emitted when using an in-place defined ``list``
-    or ``tuple`` to do a membership test. ``sets`` are better optimized for that.
-
-    Closes #4776
+  Closes #4776
 
 
 What's New in Pylint 2.10.2?
