--- tests/functional/ext/code_style/code_style_use_set_membership.rc
@@ -1,2 +1,0 @@
-[MASTER]
-load-plugins=pylint.extensions.code_style
