+++ tests/functional/ext/set_membership/set_membership_use_set_membership.rc
@@ -1,0 +1,2 @@
+[MASTER]
+load-plugins=pylint.extensions.set_membership
